Refactor: Split procedure_tests.rs into smaller focused modules
Reorganizes procedure tests from single 2,124-line file into modular structure:

**New structure:**
```
procedures/
├── mod.rs                  # Module declarations and shared utilities
├── stored_procedures.rs    # Basic CREATE/DROP/CALL tests
├── functions.rs            # Function tests
├── parameters.rs           # Parameter binding and execution tests
├── control_flow.rs         # IF/WHILE/LOOP/REPEAT control flow tests
├── error_messages.rs       # Error message validation tests
└── edge_cases.rs           # Edge cases and robustness tests
```

**Benefits:**
- ✅ Easier to locate specific tests
- ✅ Faster incremental compilation
- ✅ Smaller, more focused test modules
- ✅ Better code organization
- ✅ Easier PR reviews

Closes #2097
